+++ src/WebCompiler/Compile/LessOptions.cs
@@ -1,28 +1,49 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using Newtonsoft.Json;
 
 namespace WebCompiler
 {
-    class LessOptions : BaseOptions
+    public class LessOptions : BaseOptions<LessOptions>
     {
         private const string trueStr = "true";
 
-        public LessOptions(Config config)
-        {
-            StrictMath = GetValue(config, "strictMath").ToLowerInvariant() == trueStr;
-            StrictUnits = GetValue(config, "strictUnits").ToLowerInvariant() == trueStr;
-            RootPath = GetValue(config, "rootPath");
+        public LessOptions()
+        { }
 
-            string relativeUrls = GetValue(config, "relativeUrls").ToLowerInvariant();
-            if (relativeUrls == "false")
-                RelativeUrls = false;
+        protected override void LoadSettings(Config config)
+        {
+            var strictMath = GetValue(config, "strictMath");
+            if (strictMath != null)
+                StrictMath = strictMath.ToLowerInvariant() == trueStr;
+
+            var strictUnits = GetValue(config, "strictUnits");
+            if (strictUnits != null)
+                StrictUnits = strictUnits.ToLowerInvariant() == trueStr;
+
+            var rootPath = GetValue(config, "rootPath");
+            if (RootPath != null)
+                RootPath = RootPath;
+
+            var relativeUrls = GetValue(config, "relativeUrls");
+            if (relativeUrls != null)
+                RelativeUrls = relativeUrls.ToLowerInvariant() == trueStr;
         }
 
-        public bool StrictMath { get; set; }
+        protected override string CompilerFileName
+        {
+            get { return "less"; }
+        }
 
-        public bool StrictUnits { get; set; }
+        [JsonProperty("strictMath")]
+        public bool StrictMath { get; set; } = false;
 
+        [JsonProperty("strictUnits")]
+        public bool StrictUnits { get; set; } = false;
+
+        [JsonProperty("relativeUrls")]
         public bool RelativeUrls { get; set; } = true;
 
-        public string RootPath { get; set; }
+        [JsonProperty("rootPath")]
+        public string RootPath { get; set; } = "";
     }
 }