+++ src/WebCompilerTest/Compile/ScssOptionsTest.cs
@@ -11,7 +11,7 @@ namespace WebCompilerTest
         public void OutputStyleNested()
         {
             var configs = ConfigHandler.GetConfigs("../../artifacts/options/scss/scssconfignested.json");
-            var result = new WebCompiler.SassOptions(configs.ElementAt(0));
+            var result =  WebCompiler.SassOptions.FromConfig(configs.ElementAt(0));
             Assert.AreEqual("nested", result.OutputStyle);
         }
 
@@ -19,7 +19,7 @@ namespace WebCompilerTest
         public void OutputStyleExpanded()
         {
             var configs = ConfigHandler.GetConfigs("../../artifacts/options/scss/scssconfigexpanded.json");
-            var result = new WebCompiler.SassOptions(configs.ElementAt(0));
+            var result = WebCompiler.SassOptions.FromConfig(configs.ElementAt(0));
             Assert.AreEqual("expanded", result.OutputStyle);
         }
 
@@ -27,7 +27,7 @@ namespace WebCompilerTest
         public void OutputStyleCompact()
         {
             var configs = ConfigHandler.GetConfigs("../../artifacts/options/scss/scssconfigcompact.json");
-            var result = new WebCompiler.SassOptions(configs.ElementAt(0));
+            var result = WebCompiler.SassOptions.FromConfig(configs.ElementAt(0));
             Assert.AreEqual("compact", result.OutputStyle);
         }
 
@@ -35,23 +35,15 @@ namespace WebCompilerTest
         public void OutputStyleCompressed()
         {
             var configs = ConfigHandler.GetConfigs("../../artifacts/options/scss/scssconfigcompressed.json");
-            var result = new WebCompiler.SassOptions(configs.ElementAt(0));
+            var result = WebCompiler.SassOptions.FromConfig(configs.ElementAt(0));
             Assert.AreEqual("compressed", result.OutputStyle);
-        }
-
-        [TestMethod, TestCategory("SCSSOptions")]
-        public void OutputStyleEcho()
-        {
-            var configs = ConfigHandler.GetConfigs("../../artifacts/options/scss/scssconfigecho.json");
-            var result = new WebCompiler.SassOptions(configs.ElementAt(0));
-            Assert.AreEqual("echo", result.OutputStyle);
         }
 
         [TestMethod, TestCategory("SCSSOptions")]
         public void Precision()
         {
             var configs = ConfigHandler.GetConfigs("../../artifacts/options/scss/scssconfigprecision.json");
-            var result = new WebCompiler.SassOptions(configs.ElementAt(0));
+            var result = WebCompiler.SassOptions.FromConfig(configs.ElementAt(0));
             Assert.AreEqual(3, result.Precision);
         }
     }