+++ src/WebCompilerVsix/Constants.cs
@@ -3,6 +3,7 @@
     class Constants
     {
         public const string CONFIG_FILENAME = "compilerconfig.json";
+        public const string DEFAULTS_FILENAME = "compilerconfig.json.defaults";
         public const string VSIX_NAME = "Web Compiler";
         public const string NUGET_ID = "BuildWebCompiler";
     }