+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [x] Use `node-sass` to compile Sass files (#74)
 - [x] Source Maps support for Sass compiler
 - [x] LESS: Adjust relative paths in @import files by default
+- [x] Project wide global compiler options (#69)
 - [ ] Generate gulpfile.js from compilerconfig.json (#34)
 - [ ] Preview window (#6)
 - [ ] File globbing pattern support (#49)